+++ NEWS.md
@@ -5,6 +5,7 @@
 display a summary for individual response categories.
 * Updated `tm_t_tte` to optionally compare between arms, removed `conf_level` argument.
 * Updated `tm_g_km` to optionally compare between arms.
+* Extend `tm_g_km` to optionally scale X axis range in case of multiple plots.
 
 # teal.modules.clinical 0.8.3